requirements.txt
# Django web framework: High-level Python Web framework that encourages rapid development and clean, pragmatic design.
django==2.0.5
django-cleanup==2.1.0

pytz==2018.4
pillow==5.1.0

# Django REST framework: Powerful and flexible toolkit for building Web APIs.
djangorestframework==3.8.2
django-cors-headers==2.2.0

pyjwt==1.6.4

# Memcached: Distributed memory object caching system.
# python-memcached==1.59

# Database adapters
# psycopg2-binary==2.7.4
# mysqlclient==1.3.12

# Django channels framework: Project that takes Django and extends its abilities beyond HTTP - to handle WebSockets, chat protocols, IoT protocols, and more.
channels==2.1.0
channels-redis==2.1.1
daphne==2.1.0

# Celery: Distributed task queue. Celery is an asynchronous task queue/job queue based on distributed message passing.
celery==4.1.1
sqlalchemy==1.2.7

# Redis: In-memory data structure store, used as a database, cache and message broker.
# redis==2.10.6

# HTTP requests
requests==2.18.4

# Sentry: Error tracking that helps developers monitor and fix crashes in real time.
raven==6.8.0
